+++ src/components/popover/popover.js
@@ -235,7 +235,10 @@ export class EuiPopover extends Component {
         );
 
       setTimeout(() => {
-        this.setState({ isOpenStable: true }, this.positionPopoverFixed);
+        this.setState({ isOpenStable: true }, () => {
+          this.positionPopoverFixed();
+          this.updateFocus();
+        });
       }, durationMatch + delayMatch);
     }
 
@@ -258,8 +261,6 @@ export class EuiPopover extends Component {
         });
       }, 250);
     }
-
-    this.updateFocus();
   }
 
   componentWillUnmount() {
@@ -473,6 +474,7 @@ export class EuiPopover extends Component {
       panel = (
         <EuiPortal insert={insert}>
           <EuiFocusTrap
+            returnFocus={!this.state.isOpening} // Ignore temporary state of indecisive focus
             clickOutsideDisables={true}
             initialFocus={initialFocus}
             disabled={!ownFocus}>